Comit By Rohit

+++ SwachBharat.CMS.Bll.Sevices/AllServices/ScreenService.cs
@@ -1614,7 +1614,7 @@ namespace SwachBharat.CMS.Bll.Services
                             model.ComgcTarget = data.ComgcTarget;
                             model.userDesignation = data.userDesignation;
                             model.ZoneId = data.ZoneId;
-                            model.PrabhagId = data.PrabhagId;
+                            model.PrabhagId = (int)data.PrabhagId;
                             //model.EmployeeType = Emptype;
                             db.SaveChanges();
                         }
